When loading a kzip with old max-inclusive movement area, don’t increment it by 1.
The movement area will appear smaller by 1 compared to before. But the data is still contained in the kzip.
If we instead incremented the movement area by 1 to keep appearance consistent to before, the newly saved kzip would have a different movement area than before which might break client code.